Add containerSettings to it's own file
